+++ packages/framework/src/Model/Advert/AdvertFactory.php
@@ -2,14 +2,31 @@
 
 namespace Shopsys\FrameworkBundle\Model\Advert;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class AdvertFactory implements AdvertFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Advert\AdvertData $data
      * @return \Shopsys\FrameworkBundle\Model\Advert\Advert
      */
     public function create(AdvertData $data): Advert
     {
-        return new Advert($data);
+        $classData = $this->entityNameResolver->resolve(Advert::class);
+
+        return new $classData($data);
     }
 }