+++ packages/framework/src/Model/Customer/UserFactory.php
@@ -2,8 +2,23 @@
 
 namespace Shopsys\FrameworkBundle\Model\Customer;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class UserFactory implements UserFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Customer\UserData $userData
      * @param \Shopsys\FrameworkBundle\Model\Customer\BillingAddress $billingAddress
@@ -15,6 +30,8 @@ class UserFactory implements UserFactoryInterface
         BillingAddress $billingAddress,
         ?DeliveryAddress $deliveryAddress
     ): User {
-        return new User($userData, $billingAddress, $deliveryAddress);
+        $classData = $this->entityNameResolver->resolve(User::class);
+
+        return new $classData($userData, $billingAddress, $deliveryAddress);
     }
 }