+++ packages/framework/src/Model/Order/OrderNumberSequenceFactory.php
@@ -2,8 +2,23 @@
 
 namespace Shopsys\FrameworkBundle\Model\Order;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class OrderNumberSequenceFactory implements OrderNumberSequenceFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param int $id
      * @param string $number
@@ -11,6 +26,8 @@ class OrderNumberSequenceFactory implements OrderNumberSequenceFactoryInterface
      */
     public function create(int $id, string $number): OrderNumberSequence
     {
-        return new OrderNumberSequence($id, $number);
+        $classData = $this->entityNameResolver->resolve(OrderNumberSequence::class);
+
+        return new $classData($id, $number);
     }
 }