+++ packages/framework/src/Model/Order/Status/OrderStatusFactory.php
@@ -2,8 +2,23 @@
 
 namespace Shopsys\FrameworkBundle\Model\Order\Status;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class OrderStatusFactory implements OrderStatusFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Order\Status\OrderStatusData $data
      * @param int $type
@@ -11,6 +26,8 @@ class OrderStatusFactory implements OrderStatusFactoryInterface
      */
     public function create(OrderStatusData $data, int $type): OrderStatus
     {
-        return new OrderStatus($data, $type);
+        $classData = $this->entityNameResolver->resolve(OrderStatus::class);
+
+        return new $classData($data, $type);
     }
 }