+++ packages/framework/src/Model/PersonalData/PersonalDataAccessRequestFactory.php
@@ -2,14 +2,31 @@
 
 namespace Shopsys\FrameworkBundle\Model\PersonalData;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class PersonalDataAccessRequestFactory implements PersonalDataAccessRequestFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\PersonalData\PersonalDataAccessRequestData $data
      * @return \Shopsys\FrameworkBundle\Model\PersonalData\PersonalDataAccessRequest
      */
     public function create(PersonalDataAccessRequestData $data): PersonalDataAccessRequest
     {
-        return new PersonalDataAccessRequest($data);
+        $classData = $this->entityNameResolver->resolve(PersonalDataAccessRequest::class);
+
+        return new $classData($data);
     }
 }