+++ packages/framework/src/Model/Pricing/Vat/VatFactory.php
@@ -2,14 +2,31 @@
 
 namespace Shopsys\FrameworkBundle\Model\Pricing\Vat;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class VatFactory implements VatFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Pricing\Vat\VatData $data
      * @return \Shopsys\FrameworkBundle\Model\Pricing\Vat\Vat
      */
     public function create(VatData $data): Vat
     {
-        return new Vat($data);
+        $classData = $this->entityNameResolver->resolve(Vat::class);
+
+        return new $classData($data);
     }
 }