+++ packages/framework/src/Model/Product/Accessory/ProductAccessoryFactory.php
@@ -2,10 +2,24 @@
 
 namespace Shopsys\FrameworkBundle\Model\Product\Accessory;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
 use Shopsys\FrameworkBundle\Model\Product\Product;
 
 class ProductAccessoryFactory implements ProductAccessoryFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Product\Product $product
      * @param \Shopsys\FrameworkBundle\Model\Product\Product $accessory
@@ -17,6 +31,8 @@ class ProductAccessoryFactory implements ProductAccessoryFactoryInterface
         Product $accessory,
         int $position
     ): ProductAccessory {
-        return new ProductAccessory($product, $accessory, $position);
+        $classData = $this->entityNameResolver->resolve(ProductAccessory::class);
+
+        return new $classData($product, $accessory, $position);
     }
 }