+++ packages/framework/src/Model/Product/Brand/BrandFactory.php
@@ -2,14 +2,31 @@
 
 namespace Shopsys\FrameworkBundle\Model\Product\Brand;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class BrandFactory implements BrandFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Product\Brand\BrandData $data
      * @return \Shopsys\FrameworkBundle\Model\Product\Brand\Brand
      */
     public function create(BrandData $data): Brand
     {
-        return new Brand($data);
+        $classData = $this->entityNameResolver->resolve(Brand::class);
+
+        return new $classData($data);
     }
 }