+++ packages/framework/src/Model/Product/Parameter/ParameterFactory.php
@@ -2,14 +2,31 @@
 
 namespace Shopsys\FrameworkBundle\Model\Product\Parameter;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class ParameterFactory implements ParameterFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterData $data
      * @return \Shopsys\FrameworkBundle\Model\Product\Parameter\Parameter
      */
     public function create(ParameterData $data): Parameter
     {
-        return new Parameter($data);
+        $classData = $this->entityNameResolver->resolve(Parameter::class);
+
+        return new $classData($data);
     }
 }