+++ packages/framework/src/Model/Product/ProductFactory.php
@@ -2,15 +2,32 @@
 
 namespace Shopsys\FrameworkBundle\Model\Product;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class ProductFactory implements ProductFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Product\ProductData $data
      * @return \Shopsys\FrameworkBundle\Model\Product\Product
      */
     public function create(ProductData $data): Product
     {
-        return Product::create($data);
+        $classData = $this->entityNameResolver->resolve(Product::class);
+
+        return $classData::create($data);
     }
 
     /**
@@ -20,6 +37,8 @@ class ProductFactory implements ProductFactoryInterface
      */
     public function createMainVariant(ProductData $data, array $variants): Product
     {
-        return Product::createMainVariant($data, $variants);
+        $classData = $this->entityNameResolver->resolve(Product::class);
+
+        return $classData::createMainVariant($data, $variants);
     }
 }