+++ packages/framework/src/Model/Product/TopProduct/TopProductFactory.php
@@ -2,10 +2,24 @@
 
 namespace Shopsys\FrameworkBundle\Model\Product\TopProduct;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
 use Shopsys\FrameworkBundle\Model\Product\Product;
 
 class TopProductFactory implements TopProductFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Product\Product $product
      * @param int $domainId
@@ -17,6 +31,8 @@ class TopProductFactory implements TopProductFactoryInterface
         int $domainId,
         int $position
     ): TopProduct {
-        return new TopProduct($product, $domainId, $position);
+        $classData = $this->entityNameResolver->resolve(TopProduct::class);
+
+        return new $classData($product, $domainId, $position);
     }
 }