+++ packages/framework/src/Model/Product/Unit/UnitFactory.php
@@ -2,14 +2,31 @@
 
 namespace Shopsys\FrameworkBundle\Model\Product\Unit;
 
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
+
 class UnitFactory implements UnitFactoryInterface
 {
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    private $entityNameResolver;
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver $entityNameResolver
+     */
+    public function __construct(EntityNameResolver $entityNameResolver)
+    {
+        $this->entityNameResolver = $entityNameResolver;
+    }
+
     /**
      * @param \Shopsys\FrameworkBundle\Model\Product\Unit\UnitData $data
      * @return \Shopsys\FrameworkBundle\Model\Product\Unit\Unit
      */
     public function create(UnitData $data): Unit
     {
-        return new Unit($data);
+        $classData = $this->entityNameResolver->resolve(Unit::class);
+
+        return new $classData($data);
     }
 }